+++ packages/core/src/web/app/views/beambox/Right-Panels/Options-Blocks/TextOptions/index.module.scss
